@@ -17,6 +17,9 @@
     "test": "npx jest",
     "all": "npm run format:write && npm run lint && npm run test && npm run coverage && npm run package"
   },
+  "overrides": {
+    "eslint": "^9.7.0"
+  },
   "type": "module",
   "repository": {
     "type": "git",
@@ -80,24 +83,26 @@
     "@actions/io": "^1.1.2",
     "@actions/tool-cache": "^2.0.1",
     "semver": "^7.5.4",
-    "strip-ansi": "^7.1.0",
-    "typescript-eslint": "^7.16.1"
+    "strip-ansi": "^7.1.0"
   },
   "devDependencies": {
+    "@eslint/js": "^9.7.0",
     "@types/jest": "^29.5.12",
     "@types/node": "^20.14.9",
     "@vercel/ncc": "^0.38.1",
-    "eslint": "^8.57.0",
+    "eslint": "^9.7.0",
     "eslint-plugin-github": "^4.10.2",
     "eslint-plugin-jest": "^27.9.0",
     "eslint-plugin-jsonc": "^2.16.0",
     "eslint-plugin-prettier": "^5.1.3",
+    "globals": "^15.8.0",
     "jest": "^29.7.0",
     "js-yaml": "^4.1.0",
     "prettier": "^3.3.2",
     "prettier-eslint": "^16.3.0",
     "ts-jest": "^29.1.5",
-    "typescript": "^5.5.3"
+    "typescript": "^5.5.3",
+    "typescript-eslint": "^7.16.1"
   },
   "volta": {
     "node": "20.10.0"
