@@ -1,9 +1,7 @@
-module.exports = {
-  clearMocks: true,
-  moduleFileExtensions: ['js', 'ts'],
-  testMatch: ['**/*.test.ts'],
-  transform: {
-    '^.+\\.ts$': 'ts-jest'
-  },
-  verbose: true
+export const clearMocks = true
+export const moduleFileExtensions = ['js', 'ts']
+export const testMatch = ['**/*.test.ts']
+export const transform = {
+  '^.+\\.ts$': 'ts-jest'
 }
+export const verbose = true
