@@ -1,10 +1,17 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.QPM_COMMAND_CACHE_PATH = exports.QPM_COMMAND_PUBLISH = exports.QPM_COMMAND_RESTORE = exports.QPM_REPOSITORY_BRANCH = exports.QPM_REPOSITORY_WORKFLOW_NAME = exports.QPM_REPOSITORY_NAME = exports.QPM_REPOSITORY_OWNER = void 0;
-exports.QPM_REPOSITORY_OWNER = 'QuestPackageManager';
-exports.QPM_REPOSITORY_NAME = 'QPM.CLI';
-exports.QPM_REPOSITORY_WORKFLOW_NAME = 'cargo-build';
-exports.QPM_REPOSITORY_BRANCH = 'main';
-exports.QPM_COMMAND_RESTORE = 'restore';
-exports.QPM_COMMAND_PUBLISH = 'publish';
-exports.QPM_COMMAND_CACHE_PATH = 'cache path';
+'use strict'
+Object.defineProperty(exports, '__esModule', { value: true })
+exports.QPM_COMMAND_CACHE_PATH =
+  exports.QPM_COMMAND_PUBLISH =
+  exports.QPM_COMMAND_RESTORE =
+  exports.QPM_REPOSITORY_BRANCH =
+  exports.QPM_REPOSITORY_WORKFLOW_NAME =
+  exports.QPM_REPOSITORY_NAME =
+  exports.QPM_REPOSITORY_OWNER =
+    void 0
+exports.QPM_REPOSITORY_OWNER = 'QuestPackageManager'
+exports.QPM_REPOSITORY_NAME = 'QPM.CLI'
+exports.QPM_REPOSITORY_WORKFLOW_NAME = 'cargo-build'
+exports.QPM_REPOSITORY_BRANCH = 'main'
+exports.QPM_COMMAND_RESTORE = 'restore'
+exports.QPM_COMMAND_PUBLISH = 'publish'
+exports.QPM_COMMAND_CACHE_PATH = 'cache path'
