@@ -4,4 +4,4 @@
 import { run } from './main.js'
 
 // eslint-disable-next-line @typescript-eslint/no-floating-promises
-run()
+await run()
